gui: split receive view into module with modals submodule
