+++ liana-gui/src/app/view/receive/mod.rs
@@ -1,34 +1,26 @@
+mod modals;
+pub use modals::{qr_modal, verify_address_modal};
+
 use std::collections::{HashMap, HashSet};
 
 use iced::{
     alignment::Horizontal,
-    widget::{qr_code, Button, Space},
+    widget::{Button, Space},
     Alignment, Length,
 };
 
-use liana::miniscript::bitcoin::{
-    self,
-    bip32::{ChildNumber, Fingerprint},
-    Address,
-};
+use liana::miniscript::bitcoin;
 
 use liana_ui::{
     component::{
-        button, card, form, scrollable,
+        button, form, scrollable,
         text::{legacy, Text},
     },
     icon, theme,
     widget::*,
 };
 
-use crate::{
-    app::{
-        error::Error,
-        menu::Menu,
-        view::{hw, warning::warn},
-    },
-    hw::HardwareWallet,
-};
+use crate::app::menu::Menu;
 
 use super::message::Message;
 
@@ -229,55 +221,4 @@ pub fn receive<'a>(
         )
         .spacing(20)
         .into()
-}
-
-pub fn verify_address_modal<'a>(
-    warning: Option<&Error>,
-    hws: &'a [HardwareWallet],
-    chosen_hws: &HashSet<Fingerprint>,
-    address: &Address,
-    derivation_index: &ChildNumber,
-) -> Element<'a, Message> {
-    Column::new()
-        .push_maybe(warning.map(|w| warn(Some(w))))
-        .push(card::simple(
-            Column::new()
-                .push(
-                    Column::new()
-                        .push(liana_ui::component::panels::receive::verify_address_modal(
-                            address,
-                            derivation_index,
-                            Message::Clipboard(address.to_string()),
-                        ))
-                        .push(
-                            legacy::text("Select device to verify address on:").width(Length::Fill),
-                        )
-                        .spacing(10)
-                        .push(hws.iter().enumerate().fold(
-                            Column::new().spacing(10),
-                            |col, (i, hw)| {
-                                col.push(hw::hw_list_view_verify_address(
-                                    i,
-                                    hw,
-                                    if let HardwareWallet::Supported { fingerprint, .. } = hw {
-                                        chosen_hws.contains(fingerprint)
-                                    } else {
-                                        false
-                                    },
-                                ))
-                            },
-                        ))
-                        .width(Length::Fill),
-                )
-                .spacing(20)
-                .width(Length::Fill)
-                .align_x(Alignment::Center),
-        ))
-        .width(Length::Fill)
-        .max_width(750)
-        .into()
-}
-
-pub fn qr_modal<'a>(qr: &'a qr_code::Data, address: &'a str) -> Element<'a, Message> {
-    liana_ui::component::panels::receive::qr_display(qr, address)
 }
